refactor: 피드백 반영 - 패키지 분리
